+++ cmd/cluster/cluster.go
@@ -56,7 +56,7 @@ manage the entire lifecycle of an Astarte instance, including its installation a
 // Set here all custom resources for Astarte
 var (
 	kubernetesClient              *kubernetes.Clientset
-	kubernetesAPIExtensionsClient *apiextensions.Clientset
+	kubernetesAPIExtensionsClient *apiextensions.Clientset //nolint
 	kubernetesDynamicClient       dynamic.Interface
 
 	astarteV1Alpha1 = schema.GroupVersionResource{
@@ -79,16 +79,7 @@ var (
 		Version:  "v2alpha1",
 		Resource: "astartes",
 	}
-	aviV1Alpha1 = schema.GroupVersionResource{
-		Group:    "api.astarte-platform.org",
-		Version:  "v1alpha1",
-		Resource: "astartevoyageringresses",
-	}
-	adiV1Alpha1 = schema.GroupVersionResource{
-		Group:    "ingress.astarte-platform.org",
-		Version:  "v1alpha1",
-		Resource: "astartedefaultingresses",
-	}
+
 	astarteCRVersions = map[string]schema.GroupVersionResource{
 		"v1alpha1": astarteV1Alpha1,
 		"v1alpha2": astarteV1Alpha2,